Extract MessageRow to whole-file platform gate for CI boundary check
The multiplatform boundary checker flags inline #if os(macOS) blocks in
non-Platform files. MessageRow contains macOS-specific quote-reply state
and popover logic. Moving it to its own file with a whole-file platform
gate (#if os(macOS) / #else / #endif) satisfies the checker while
keeping the cross-platform fallback.
